move common logic from fragments to base fragment
