+++ presentation/src/main/res/layout/dialogue_city.xml
@@ -41,29 +41,20 @@
         <Button
             android:id="@+id/btnGps"
             style="@style/DialogueBtnStyle"
-            android:layout_width="0dp"
             android:padding="@dimen/offset_10"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
             android:drawableStart="@drawable/ic_add_location"
             android:text="@string/dialogue_btn_gps" />
 
         <Button
             android:id="@+id/btnDate"
             style="@style/DialogueBtnStyle"
-            android:layout_width="0dp"
             android:padding="@dimen/offset_10"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
             android:drawableStart="@drawable/ic_month"
             android:text="@string/home_city_date" />
 
         <Button
             android:id="@+id/buttonSave"
             style="@style/DialogueBtnStyle"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
             android:text="@string/btn_save" />
 
     </LinearLayout>
