+++ presentation/src/main/res/layout/layout_app_bar.xml
@@ -18,7 +18,7 @@
         </FrameLayout>
 
         <androidx.cardview.widget.CardView
-            android:id="@+id/cvAvatar"
+            android:id="@+id/cv_avatar"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:visibility="gone"
